\name{maize.G}
\alias{maize.G}
\title{Genomic relationship matrix for maize lines}
\description{
A matrix with relationship between individuals for 
parents of two heterotic groups. The matrix was computed 
using 511 SNPs using the function A.mat included in 
rrBLUP package (Endelman, 2011). The row names and column 
names of this matrix corresponds to the genotype 
identifiers for Parent 1 and Parent 2.
}

\keyword{datasets}
